Add `xhttp.IsTokenChar` function
